Fastjet cmake and benchmark fixes (#13)
* Update fastjet-finder CMake

CMakeLists file now is adapted for modern Fastjet versions that are
built with CMake (>=3.5.0). Old FindFastJet.cmake is removed.

* Fix argument name in README

* Consistency update for fastjet-finder

Align the arguments for fastjet-finder to be the same as those for the
Julia test scripts, e.g., -m or --nsamples for the number of trial
repeats.

Ensure that algorithm and power values from benchmark.jl are correctly
passed to fastjet-finder. It seems the algorithm setting was actually
being _ignored_, meaning only pp-type algorithms actually got run.

* Set fastjet output to "debug"

* Update antikt-python to new tag